# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ByteBite food ordering and delivery API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: menu
    description: Menu and restaurant operations
  - name: orders
    description: Order management
  - name: drivers
    description: Driver operations
  - name: checkout
    description: Payment and checkout

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /restaurants:
    get:
      operationId: listRestaurants
      tags: [menu]
      summary: List all restaurants
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"

  /restaurants/mine:
    get:
      operationId: listMyRestaurants
      tags: [menu]
      summary: List restaurants owned by current user
      responses:
        "200":
          description: List of owned restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"

  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [menu]
      summary: Get restaurant with menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant with menu items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantWithMenu"
        "404":
          description: Restaurant not found
    patch:
      operationId: updateRestaurant
      tags: [menu]
      summary: Update a restaurant owned by the current user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantUpdate"
      responses:
        "200":
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "403":
          description: Not the owner
        "404":
          description: Restaurant not found
    delete:
      operationId: deleteRestaurant
      tags: [menu]
      summary: Delete a restaurant owned by the current user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "403":
          description: Not the owner
        "404":
          description: Restaurant not found

  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      parameters:
        - name: restaurantId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /menu/items:
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Add a menu item to a restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "403":
          description: Not the restaurant owner
        "400":
          description: Invalid request

  /menu/items/{id}:
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "403":
          description: Not the restaurant owner
        "404":
          description: Menu item not found
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "403":
          description: Not the restaurant owner
        "404":
          description: Menu item not found

  /menu/categories:
    get:
      operationId: listMenuCategories
      tags: [menu]
      summary: List menu categories
      parameters:
        - name: restaurantId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of category names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [customer, restaurant, driver]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found

  /orders/summary:
    get:
      operationId: getOrdersSummary
      tags: [orders]
      summary: Get orders summary stats
      responses:
        "200":
          description: Order statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummary"

  /drivers/available:
    get:
      operationId: listAvailableOrders
      tags: [drivers]
      summary: List orders available for pickup
      responses:
        "200":
          description: Orders ready for driver pickup
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /drivers/my-orders:
    get:
      operationId: listDriverOrders
      tags: [drivers]
      summary: List orders assigned to a driver
      parameters:
        - name: driverName
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Driver's orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /drivers/{orderId}/accept:
    patch:
      operationId: acceptOrder
      tags: [drivers]
      summary: Driver accepts an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverAcceptInput"
      responses:
        "200":
          description: Order accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found

  /checkout/session:
    post:
      operationId: createCheckoutSession
      tags: [checkout]
      summary: Create Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutSessionInput"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "400":
          description: Invalid request

  /checkout/webhook:
    post:
      operationId: stripeWebhook
      tags: [checkout]
      summary: Stripe webhook handler
      responses:
        "200":
          description: Webhook processed

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Restaurant:
      type: object
      required: [id, name, description, imageUrl, cuisineType, deliveryTime, rating]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        cuisineType:
          type: string
        deliveryTime:
          type: integer
        rating:
          type: number

    MenuItem:
      type: object
      required: [id, restaurantId, name, description, price, category, imageUrl, available]
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        available:
          type: boolean

    RestaurantWithMenu:
      type: object
      required: [id, name, description, imageUrl, cuisineType, deliveryTime, rating, menuItems]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        cuisineType:
          type: string
        deliveryTime:
          type: integer
        rating:
          type: number
        menuItems:
          type: array
          items:
            $ref: "#/components/schemas/MenuItem"

    OrderItem:
      type: object
      required: [menuItemId, name, price, quantity]
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer

    Order:
      type: object
      required: [id, customerName, restaurantId, restaurantName, status, total, items, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        restaurantId:
          type: integer
        restaurantName:
          type: string
        driverName:
          type: ["string", "null"]
        status:
          type: string
          enum: [placed, accepted, cooking, ready, delivered]
        total:
          type: number
        deliveryAddress:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        stripeSessionId:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [pending, paid, failed]
        createdAt:
          type: string

    OrderInput:
      type: object
      required: [customerName, restaurantId, restaurantName, total, items, deliveryAddress]
      properties:
        customerName:
          type: string
        restaurantId:
          type: integer
        restaurantName:
          type: string
        total:
          type: number
        deliveryAddress:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [placed, accepted, cooking, ready, delivered]

    OrderSummary:
      type: object
      required: [total, placed, accepted, cooking, ready, delivered, revenue]
      properties:
        total:
          type: integer
        placed:
          type: integer
        accepted:
          type: integer
        cooking:
          type: integer
        ready:
          type: integer
        delivered:
          type: integer
        revenue:
          type: number

    DriverAcceptInput:
      type: object
      required: [driverName]
      properties:
        driverName:
          type: string

    CheckoutSessionInput:
      type: object
      required: [orderId, items]
      properties:
        orderId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    MenuItemInput:
      type: object
      required: [restaurantId, name, description, price, category, imageUrl]
      properties:
        restaurantId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        available:
          type: boolean

    MenuItemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: string
        available:
          type: boolean

    RestaurantUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        cuisineType:
          type: string
        deliveryTime:
          type: integer

    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    CheckoutSession:
      type: object
      required: [sessionId, url]
      properties:
        sessionId:
          type: string
        url:
          type: string